# Relieve db pressure on api/health check