--- a/src/metabase/db/connection_pool_setup.clj
+++ b/src/metabase/db/connection_pool_setup.clj
@@ -6,7 +6,7 @@
             [schema.core :as s])
   (:import [com.mchange.v2.c3p0 ConnectionCustomizer PoolBackedDataSource]))
 
-(def ^:private latest-checkin (atom nil))
+(def ^:private latest-activity (atom nil))
 
 (def ^:private ^java.time.Duration recent-window-duration (t/seconds 15))
 
@@ -16,17 +16,20 @@
     (t/after? activity (t/minus (t/offset-date-time) duration))))
 
 (defn recent-activity?
-  "Returns true if there has been recent activity. Define recent activity as an application db connection checked in
-  within 15 seconds. Check-in means a query succeeded and the db connection is no longer needed."
+  "Returns true if there has been recent activity. Define recent activity as an application db connection checked in,
+  checked out, or acquired within [[recent-window-duration]]. Check-in means a query succeeded and the db connection
+  is no longer needed."
   []
-  (recent-activity?* @latest-checkin recent-window-duration))
+  (recent-activity?* @latest-activity recent-window-duration))
 
-(defrecord CheckinTracker []
+(defrecord DbActivityTracker []
   ConnectionCustomizer
-  (onAcquire [_ _connection _identity-token])
+  (onAcquire [_ _connection _identity-token]
+    (reset! latest-activity (t/offset-date-time)))
   (onCheckIn [_ _connection _identity-token]
-    (reset! latest-checkin (t/offset-date-time)))
-  (onCheckOut [_ _connection _identity-token])
+    (reset! latest-activity (t/offset-date-time)))
+  (onCheckOut [_ _connection _identity-token]
+    (reset! latest-activity (t/offset-date-time)))
   (onDestroy [_ _connection _identity-token]))
 
 (defn- register-customizer!
@@ -42,7 +45,7 @@
     (.put ^java.util.HashMap (.get field com.mchange.v2.c3p0.C3P0Registry)
           (.getName klass) (.newInstance klass))))
 
-(register-customizer! CheckinTracker)
+(register-customizer! DbActivityTracker)
 
 (def ^:private application-db-connection-pool-props
   "Options for c3p0 connection pool for the application DB. These are set in code instead of a properties file because
@@ -51,7 +54,7 @@
   below (jump to the 'Simple advice on Connection testing' section.)"
   (merge
    {"idleConnectionTestPeriod" 60
-    "connectionCustomizerClassName" (.getName CheckinTracker)}
+    "connectionCustomizerClassName" (.getName DbActivityTracker)}
    ;; only merge in `max-pool-size` if it's actually set, this way it doesn't override any things that may have been
    ;; set in `c3p0.properties`
    (when-let [max-pool-size (config/config-int :mb-application-db-max-connection-pool-size)]