--- a/test/metabase/db/connection_pool_setup_test.clj
+++ b/test/metabase/db/connection_pool_setup_test.clj
@@ -1,11 +1,15 @@
 (ns metabase.db.connection-pool-setup-test
   (:require [clojure.java.jdbc :as jdbc]
             [clojure.test :refer :all]
+            [java-time :as t]
             [metabase.connection-pool :as connection-pool]
             [metabase.db.connection-pool-setup :as mdb.connection-pool-setup]
             [metabase.db.data-source :as mdb.data-source]
-            [metabase.test :as mt])
-  (:import com.mchange.v2.c3p0.PoolBackedDataSource))
+            [metabase.models :refer [Database]]
+            [metabase.test :as mt]
+            [toucan.db :as db])
+  (:import [com.mchange.v2.c3p0 C3P0Registry ConnectionCustomizer PoolBackedDataSource]
+           metabase.db.connection_pool_setup.CheckinTracker))
 
 (deftest connection-pool-spec-test
   (testing "Should be able to create a connection pool"
@@ -28,3 +32,44 @@
       (testing "from a connection URL"
         (test* :h2 (mdb.data-source/raw-connection-string->DataSource
                     (format "jdbc:h2:mem:%s;DB_CLOSE_DELAY=10" (mt/random-name))))))))
+
+(deftest CheckinTracker-test
+  (testing "connection customizer is registered"
+    (let [customizer (C3P0Registry/getConnectionCustomizer (.getName CheckinTracker))]
+      (is (some? customizer) "ConnectionCustomizer is not registered with c3p0")
+      (is (instance? ConnectionCustomizer customizer)
+          "checkin tracker must satisfy the c3p0 ConnectionCustomizer interface")
+      (is (instance? CheckinTracker customizer)
+          "ConnectionCustomizer is not an instance of our CheckinTracker")))
+  (testing "db activity resets counter"
+    (reset! (var-get #'mdb.connection-pool-setup/latest-checkin) nil)
+    (db/count Database) ;; trigger db access which should reset the latest checkin
+    (let [recent-checkin (deref (var-get #'mdb.connection-pool-setup/latest-checkin))]
+      (is (some? recent-checkin)
+          "Database activity did not reset latest-checkin")
+      (is (instance? java.time.temporal.Temporal recent-checkin)
+          "recent-checkin should be a temporal type (OffsetDateTime)"))))
+
+(deftest recent-activity-test
+  (testing "If latest-checkin is null"
+    (reset! (var-get #'mdb.connection-pool-setup/latest-checkin) nil)
+    (is (not (mdb.connection-pool-setup/recent-activity?)))
+    (db/count Database)
+    (testing "db activity makes `recent-activity?` true"
+      (is (mdb.connection-pool-setup/recent-activity?))))
+  (testing "If latest-checkin is stale"
+    (let [duration (var-get #'mdb.connection-pool-setup/recent-window-duration)
+          twice-duration (t/minus (t/offset-date-time) duration duration)]
+      (reset! (var-get #'mdb.connection-pool-setup/latest-checkin) twice-duration)
+      (is (not (mdb.connection-pool-setup/recent-activity?)))
+      (db/count Database)
+      (testing "db activity makes `recent-activity?` true"
+        (is (mdb.connection-pool-setup/recent-activity?)))))
+  (testing "Goes stale"
+    (with-redefs [mdb.connection-pool-setup/recent-window-duration (t/millis 30)]
+      (db/count Database)
+      (is (mdb.connection-pool-setup/recent-activity?))
+      (testing "When duration elapses should report no recent-activity"
+        (Thread/sleep 60)
+        (is (not (mdb.connection-pool-setup/recent-activity?))
+            "recent-window-duration has elapsed but still recent")))))